# Improve sift_down performance in BinaryHeap